% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueField-package.R
\docType{package}
\name{TissueField-package}
\alias{TissueField}
\alias{TissueField-package}
\title{TissueField: Steady-State Molecular Concentration Fields for Spatial Transcriptomics}
\description{
Computes continuous steady-state molecular concentration fields from discrete
mRNA transcript or protein detection coordinates using a physically motivated
diffusion-clearance model. The screened Poisson PDE is solved numerically
using one of three methods: finite-difference sparse linear system (\code{"fd"}),
Green's function FFT convolution (\code{"green"}), or Gaussian kernel density
(\code{"kde"}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RaredonLab/TissueField}
  \item Report bugs at \url{https://github.com/RaredonLab/TissueField/issues}
}

}
\author{
\strong{Maintainer}: Raredon Laboratory \email{raredon@yale.edu}

}
\keyword{internal}
